Bug fix for newlines
The newline characters are not stripped from lines of input. This causes the rightmost column to be non-empty. If the rightmost column happens to be the PEK index, the string is not empty (because it contains \n or \r\n) and it throws "[!] Warning! Multiple records with PEK entry!" and replaces the PEK key with newlines. This change simply strips the newlines